Move camel-seda and camel-file into the same class
