+++ tests/camel-integration-test/src/main/java/org/apache/karaf/camel/itests/AbstractCamelSingleComponentResultMockBasedRoute.java
@@ -30,12 +30,12 @@ public abstract class AbstractCamelSingleComponentResultMockBasedRoute implement
 
     @Override
     public CamelContext getContext() {
-        return provider.getContext(getClass());
+        return provider.getContext();
     }
 
     @Override
     public ProducerTemplate getTemplate() {
-        return provider.getTemplate(getClass());
+        return provider.getTemplate();
     }
 
     public void testRoutes() throws Exception {
