+++ tests/camel-integration-test/src/main/java/org/apache/karaf/camel/itests/CamelContextProvider.java
@@ -12,8 +12,23 @@ public interface CamelContextProvider {
     CamelContext getContext(Class<?> clazz);
 
     /**
+     * Returns the {@link CamelContext} associated with the current class according to the annotation
+     * {@link CamelKarafTestHint}.
+     */
+    default CamelContext getContext() {
+        return getContext(getClass());
+    }
+    /**
      * Returns the {@link ProducerTemplate} associated with the given class according to the annotation
      * {@link CamelKarafTestHint}.
      */
     ProducerTemplate getTemplate(Class<?> clazz);
+
+    /**
+     * Returns the {@link ProducerTemplate} associated with the current class according to the annotation
+     * {@link CamelKarafTestHint}.
+     */
+    default ProducerTemplate getTemplate() {
+        return getTemplate(getClass());
+    }
 }
